requirements.txt
slack-bolt==1.21.3
slack-sdk==3.33.5
anthropic==0.42.0
google-cloud-bigquery==3.27.0
python-dotenv==1.0.1
google-genai==0.3.0
